requirements.txt
absl-py==1.0.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
arch==5.0.1
asttokens==2.2.1
attrs==21.4.0
backcall==0.2.0
black==20.8b1
brax==0.0.13
CacheControl==0.12.11
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
chex==0.1.3
click==8.0.3
cloudpickle==1.6.0
comet-ml==3.23.1
configobj==5.0.6
configparser==5.0.2
contextlib2==21.6.0
crashtest==0.4.1
cryptography==39.0.1
cycler==0.11.0
cython==0.29.28
dataclasses==0.6
decorator==5.1.1
distlib==0.3.6
dm-control==0.0.403778684
dm-env==1.5
dm-tree==0.1.6
docker-pycreds==0.4.0
docstring-parser==0.14.1
dulwich==0.21.3
everett==2.0.1
executing==1.2.0
fasteners==0.17.3
filelock==3.9.0
flatbuffers==2.0
flax==0.3.5
frozendict==2.3.5
future==0.18.2
gast==0.5.3
gitdb==4.0.9
gitpython==3.1.24
glfw==1.12.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
gtimer==1.0.0b5
gym==0.21.0
gym-notices==0.0.6
h5py==3.4.0
html5lib==1.1
hydra-core @ https://github.com/facebookresearch/hydra/archive/8c038c62d2336688cf65da105864e01d4e6af407.tar.gz
hydra-submitit-launcher==1.1.6
idna==3.3
imageio==2.16.1
imageio-ffmpeg==0.4.2
importlib-metadata==4.13.0
installer==0.6.0
ipython==8.8.0
jaraco.classes==3.2.3
jax==0.3.1
https://storage.googleapis.com/jax-releases/cuda11/jaxlib-0.1.74+cuda11.cudnn82-cp39-none-manylinux2010_x86_64.whl
jedi==0.18.2
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
kiwisolver==1.3.1
labmaze==1.0.5
lockfile==0.12.2
lxml==4.8.0
markdown==3.3.6
markdown-it-py==2.2.0
matplotlib==3.4.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-collections==0.1.0
more-itertools==9.1.0
msgpack==1.0.2
mujoco-py==2.1.2.14
mypy==0.910
mypy-extensions==0.4.3
numpy==1.21.2
nvidia-ml-py==12.535.77
nvidia-ml-py3==7.352.0
nvitop==1.2.0
oauthlib==3.2.0
omegaconf==2.1.1
opt-einsum==3.3.0
optax==0.1.2
packaging==23.0
pandas==1.3.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.0.1
pkginfo==1.9.6
platformdirs==2.6.2
plotly==5.5.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
promise==2.3
prompt-toolkit==3.0.36
property-cached==1.6.4
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygame==2.1.2
pyglet==1.5.0
Pygments==2.14.0
PyOpenGL==3.1.6
pyparsing==3.0.6
pyproject-hooks==1.0.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytinyrenderer==0.0.13
pytz==2021.3
pyyaml==6.0
rapidfuzz==2.13.7
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==13.3.5
rliable==1.0.7
rsa==4.8
scipy==1.7.1
seaborn==0.11.2
SecretStorage==3.3.3
semantic-version==2.8.5
sentry-sdk==1.5.0
shellingham==1.5.0.post1
shortuuid==1.0.1
shtab==1.6.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
statsmodels==0.13.1
submitit==1.4.1
subprocess32==3.5.4
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
tensorflow-probability==0.16.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.11.2
tornado==6.1
tqdm==4.60.0
traitlets==5.1.1
trimesh==3.11.2
trove-classifiers==2023.3.9
typed-ast==1.4.3
typing-extensions==4.5.0
tyro==0.5.2
urllib3==1.26.7
virtualenv==20.20.0
wandb==0.12.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
werkzeug==2.0.2
wrapt==1.12.1
wurlitzer==3.0.2
yaspin==2.1.0
zipp==3.6.0
